Move configure_logger and configure_sentry outside of Config class
