+++ webapp/__init__.py
@@ -1,6 +1,5 @@
 """lambdas package."""
 
 from webapp.app import create_app
-from webapp.config import Config
 
-__all__ = ["Config", "create_app"]
+__all__ = ["create_app"]
